Refactoring of WalkAndRun, Jump, Step, Dash, GroundCheck
