+++ Assets/Data/PlayerMovementData.asset
@@ -1,0 +1,53 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f8c0a2026e03bec48954ac4d624d395a, type: 3}
+  m_Name: PlayerMovementData
+  m_EditorClassIdentifier: 
+  _cameraObject: {fileID: 0}
+  _inputHandler: {fileID: 0}
+  _characterTransform: {fileID: 0}
+  _animatorHandler: {fileID: 0}
+  _normalCamera: {fileID: 0}
+  _controller: {fileID: 0}
+  _cameraHandler: {fileID: 0}
+  _moveDirection: {x: 0, y: 0, z: 0}
+  _movementSpeed: 8
+  _rotationSpeed: 15
+  _jumpHeight: 2
+  _gravity: -15
+  _sprintSpeed: 12
+  _speedModulation: 0
+  _isGrounded: 0
+  _groundLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  _groundedOffset: -0.08
+  _didSecondJump: 0
+  _verticalVelocity: 0
+  _terminalVelocity: 50
+  _horizontalVelocity: {x: 0, y: 0, z: 0}
+  _isStepping: 0
+  _stepPower: 5
+  _stepStartupFrameNumber: 4
+  _stepActiveFrameNumber: 15
+  _stepRecoveryFrameNumber: 10
+  _isRecoveringFromStep: 0
+  _stepFrameCount: 1
+  _isStepFrameCountStarted: 0
+  _stepMovementInput: {x: 0, y: 0}
+  _currentStepMovementInput: {x: 0, y: 0}
+  _isDashing: 0
+  _dashVelocity: {x: 0, y: 0, z: 0}
+  _dashPower: 6
+  _dashStartupFrameNumber: 8
+  _dashFrameCount: 1
+  _isDashFrameCountStarted: 0
