+++ Assets/Data/PlayerMovementData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e598cc9aec7faf045b0e0ce2061e1ebb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
