+++ Assets/Data/PlayerMovementData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8c0a2026e03bec48954ac4d624d395a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
