+++ Assets/Scripts/Movements/WalkAndRun.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using UnityEditor;
 using UnityEngine;
 
 namespace DNA
@@ -7,6 +8,15 @@ namespace DNA
     public class WalkAndRun : MonoBehaviour
     {
         private PlayerMovementData _movementData;
+        private PlayerMovement _playerMovement;
+        private string _scriptableObjectPath;
+
+        private void Start()
+        {
+            _playerMovement = GetComponentInParent<PlayerMovement>();
+            _scriptableObjectPath = _playerMovement.getPlayerMovementDataPath();
+            _movementData = AssetDatabase.LoadAssetAtPath<PlayerMovementData>(_scriptableObjectPath);
+        }
 
         /// <summary>
         /// Moves the character forward, back, right, left
